Add default dirs
